melhorias no sisaq

+++ poc/sisaq/src/main/java/br/com/mineradora/dto/AquisicaoDTO.java
@@ -24,7 +24,7 @@ public class AquisicaoDTO implements Serializable {
 	
 	private BigInteger id;
 	
-	private Long quantidade;
+	private BigInteger quantidade;
 	
 	private LocalDateTime dataInclusao;
